requirements.txt
streamlit
pandas
plotly
scikit-learn
langchain-google-genai
google-generativeai
sentence-transformers
python-dotenv
numpy
